refactor: extract useRemoveAccountAlert hook, fix TransferSuccessScreen
Extract duplicated alert+factoryReset logic into shared useRemoveAccountAlert
hook. Apply to Account, AccountExpired, Settings, and TransferSuccessScreen
(which also navigated to the stuck RemoveAccountConfirmation screen).
