+++ app/src/bcsc-theme/features/account-transfer/transferer/TransferSuccessScreen.test.tsx
@@ -1,9 +1,10 @@
+import { BCSCLoadingProvider } from '@/bcsc-theme/contexts/BCSCLoadingContext'
 import { testIdWithKey } from '@bifold/core'
 import { useNavigation } from '@mocks/@react-navigation/native'
 import { BasicAppContext } from '@mocks/helpers/app'
 import { act, fireEvent, render } from '@testing-library/react-native'
 import React from 'react'
-import { BCSCScreens, BCSCStacks } from '../../../types/navigators'
+import { BCSCStacks, BCSCScreens } from '../../../types/navigators'
 import TransferSuccessScreen from './TransferSuccessScreen'
 
 describe('TransferSuccess', () => {
@@ -23,7 +24,9 @@ describe('TransferSuccess', () => {
     it('renders correctly', () => {
       const tree = render(
         <BasicAppContext>
-          <TransferSuccessScreen />
+          <BCSCLoadingProvider>
+            <TransferSuccessScreen />
+          </BCSCLoadingProvider>
         </BasicAppContext>
       )
 
@@ -35,7 +38,9 @@ describe('TransferSuccess', () => {
     it('navigates to Home when primary button is pressed', () => {
       const { getByTestId } = render(
         <BasicAppContext>
-          <TransferSuccessScreen />
+          <BCSCLoadingProvider>
+            <TransferSuccessScreen />
+          </BCSCLoadingProvider>
         </BasicAppContext>
       )
 
@@ -45,21 +50,6 @@ describe('TransferSuccess', () => {
       })
 
       expect(mockNavigation.navigate).toHaveBeenCalledWith(BCSCStacks.Tab, { screen: BCSCScreens.Home })
-    })
-
-    it('navigates to RemoveAccountConfirmation when remove account button is pressed', () => {
-      const { getByTestId } = render(
-        <BasicAppContext>
-          <TransferSuccessScreen />
-        </BasicAppContext>
-      )
-
-      const removeAccountButton = getByTestId(testIdWithKey('BCSC.Account.RemoveAccount'))
-      act(() => {
-        fireEvent.press(removeAccountButton)
-      })
-
-      expect(mockNavigation.navigate).toHaveBeenCalledWith(BCSCScreens.RemoveAccountConfirmation)
     })
   })
 })
