+++ app/src/bcsc-theme/features/account-transfer/transferer/TransferSuccessScreen.tsx
@@ -1,4 +1,5 @@
 import StatusDetails from '@/bcsc-theme/components/StatusDetails'
+import { useRemoveAccountAlert } from '@/bcsc-theme/hooks/useRemoveAccountAlert'
 import { BCSCMainStackParams, BCSCScreens, BCSCStacks } from '@/bcsc-theme/types/navigators'
 import { Button, ButtonType, ScreenWrapper, testIdWithKey } from '@bifold/core'
 import { useNavigation } from '@react-navigation/native'
@@ -10,6 +11,7 @@ import { StyleSheet } from 'react-native'
 const TransferSuccessScreen: React.FC = () => {
   const navigation = useNavigation<StackNavigationProp<BCSCMainStackParams>>()
   const { t } = useTranslation()
+  const showRemoveAccountAlert = useRemoveAccountAlert()
 
   const styles = StyleSheet.create({
     contentContainer: {
@@ -31,9 +33,7 @@ const TransferSuccessScreen: React.FC = () => {
         testID={testIdWithKey(t('BCSC.Account.RemoveAccount'))}
         buttonType={ButtonType.Critical}
         title={t('BCSC.Account.RemoveAccount')}
-        onPress={() => {
-          navigation.navigate(BCSCScreens.RemoveAccountConfirmation)
-        }}
+        onPress={showRemoveAccountAlert}
       />
     </>
   )
