+++ app/src/bcsc-theme/features/account/AccountExpiredScreen.tsx
@@ -1,15 +1,12 @@
-import { useFactoryReset } from '@/bcsc-theme/api/hooks/useFactoryReset'
 import { AppBannerSection, BCSCBanner } from '@/bcsc-theme/components/AppBanner'
 import { CardButton } from '@/bcsc-theme/components/CardButton'
 import { GENERIC_CARD_SIZE_SMALL } from '@/bcsc-theme/components/GenericCardImage'
 import { BCSCAccountContext } from '@/bcsc-theme/contexts/BCSCAccountContext'
-import { useLoadingScreen } from '@/bcsc-theme/contexts/BCSCLoadingContext'
+import { useRemoveAccountAlert } from '@/bcsc-theme/hooks/useRemoveAccountAlert'
 import { BCSCMainStackParams, BCSCScreens } from '@/bcsc-theme/types/navigators'
-import { useErrorAlert } from '@/contexts/ErrorAlertContext'
-import { AppEventCode } from '@/events/appEventCode'
-import { ScreenWrapper, TOKENS, useServices, useTheme } from '@bifold/core'
+import { ScreenWrapper, useTheme } from '@bifold/core'
 import { StackNavigationProp } from '@react-navigation/stack'
-import { useCallback, useContext } from 'react'
+import { useContext } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ScrollView, StyleSheet, View } from 'react-native'
 import HomeHeader from '../home/components/HomeHeader'
@@ -27,10 +24,7 @@ export const AccountExpiredScreen = ({ navigation }: AccountExpiredScreenProps):
   const { t } = useTranslation()
   const { Spacing } = useTheme()
   const accountContext = useContext(BCSCAccountContext)
-  const factoryReset = useFactoryReset()
-  const { emitAlert } = useErrorAlert()
-  const loadingScreen = useLoadingScreen()
-  const [logger] = useServices([TOKENS.UTIL_LOGGER])
+  const showRemoveAccountAlert = useRemoveAccountAlert()
 
   const styles = StyleSheet.create({
     scrollContainer: {
@@ -41,39 +35,6 @@ export const AccountExpiredScreen = ({ navigation }: AccountExpiredScreenProps):
       gap: Spacing.md,
     },
   })
-
-  const handleRemoveAccount = useCallback(() => {
-    emitAlert(t('Alerts.RemoveAccount.Title'), t('Alerts.RemoveAccount.Description'), {
-      event: AppEventCode.REMOVE_ACCOUNT,
-      actions: [
-        {
-          text: t('Global.Cancel'),
-          style: 'cancel',
-        },
-        {
-          text: t('Alerts.RemoveAccount.Action1'),
-          style: 'destructive',
-          onPress: async () => {
-            try {
-              loadingScreen.startLoading(t('BCSC.Account.RemoveAccountLoading'))
-
-              logger.info('[RemoveAccount] User confirmed account removal, proceeding with verification reset')
-
-              const result = await factoryReset()
-
-              if (!result.success) {
-                logger.error('[RemoveAccount] Failed to remove account', result.error)
-              }
-            } catch (error) {
-              logger.error('[RemoveAccount] Error during account removal', error as Error)
-            } finally {
-              loadingScreen.stopLoading()
-            }
-          },
-        },
-      ],
-    })
-  }, [emitAlert, t, factoryReset, loadingScreen, logger])
 
   return (
     <ScreenWrapper padded={false} scrollable={false}>
@@ -102,7 +63,7 @@ export const AccountExpiredScreen = ({ navigation }: AccountExpiredScreenProps):
           />
           <CardButton
             title={t('BCSC.AccountExpired.RemoveButton')}
-            onPress={handleRemoveAccount}
+            onPress={showRemoveAccountAlert}
           />
         </View>
       </ScrollView>
