+++ app/src/bcsc-theme/features/settings/SettingsContent.tsx
@@ -1,9 +1,6 @@
-import { useFactoryReset } from '@/bcsc-theme/api/hooks/useFactoryReset'
-import { useLoadingScreen } from '@/bcsc-theme/contexts/BCSCLoadingContext'
 import useSecureActions from '@/bcsc-theme/hooks/useSecureActions'
+import { useRemoveAccountAlert } from '@/bcsc-theme/hooks/useRemoveAccountAlert'
 import { ACCESSIBILITY_URL, ANALYTICS_URL, FEEDBACK_URL, TERMS_OF_USE_URL } from '@/constants'
-import { useErrorAlert } from '@/contexts/ErrorAlertContext'
-import { AppEventCode } from '@/events/appEventCode'
 import { BCDispatchAction, BCState } from '@/store'
 import { Analytics } from '@/utils/analytics/analytics-singleton'
 import TabScreenWrapper from '@bcsc-theme/components/TabScreenWrapper'
@@ -50,9 +47,7 @@ export const SettingsContent: React.FC<SettingsContentProps> = ({
   const { logout } = useSecureActions()
   const [logger] = useServices([TOKENS.UTIL_LOGGER])
   const [accountSecurityMethod, setAccountSecurityMethod] = useState<AccountSecurityMethod>()
-  const factoryReset = useFactoryReset()
-  const { emitAlert } = useErrorAlert()
-  const loadingScreen = useLoadingScreen()
+  const showRemoveAccountAlert = useRemoveAccountAlert()
 
   const styles = StyleSheet.create({
     container: {
@@ -152,39 +147,6 @@ export const SettingsContent: React.FC<SettingsContentProps> = ({
     }
   }
 
-  const onPressRemoveAccount = () => {
-    emitAlert(t('Alerts.CancelMobileCardSetup.Title'), t('Alerts.CancelMobileCardSetup.Description'), {
-      event: AppEventCode.CANCEL_MOBILE_CARD_SETUP,
-      actions: [
-        {
-          text: t('Global.Cancel'),
-          style: 'cancel',
-        },
-        {
-          text: t('Alerts.CancelMobileCardSetup.Action1'),
-          style: 'destructive',
-          onPress: async () => {
-            try {
-              loadingScreen.startLoading(t('BCSC.Account.RemoveAccountLoading'))
-
-              logger.info('[RemoveAccount] User confirmed account removal, proceeding with verification reset')
-
-              const result = await factoryReset()
-
-              if (!result.success) {
-                logger.error('[RemoveAccount] Failed to remove account', result.error)
-              }
-            } catch (error) {
-              logger.error('[RemoveAccount] Error during account removal', error as Error)
-            } finally {
-              loadingScreen.stopLoading()
-            }
-          },
-        },
-      ],
-    })
-  }
-
   // Pre-compute conditional values
   const isAuthenticated = store.authentication.didAuthenticate
   const showChangePIN = accountSecurityMethod !== AccountSecurityMethod.DeviceAuth && onChangePIN
@@ -240,7 +202,7 @@ export const SettingsContent: React.FC<SettingsContentProps> = ({
 
               <SettingsActionCard
                 title={t('BCSC.Settings.RemoveAccount')}
-                onPress={onPressRemoveAccount}
+                onPress={showRemoveAccountAlert}
                 textStyle={{ color: ColorPalette.semantic.error }}
               />
             </View>
